+++ Pilar_Facilitis_Sistem/Pilar_Facilitis.Domain/Entities/Servicos.cs
@@ -17,6 +17,6 @@ namespace Pilar_Facilitis.Domain.Entities
 
         public string Desc_Servicos { get; set; }
 
-        public decimal Area { get; set; }
+        public float Area { get; set; }
     }
 }
